requirements.txt
streamlit==1.40.2
openai==1.58.1
python-dotenv==1.0.1
langchain==0.3.14
langchain-openai==0.2.14
langgraph==0.2.66
pydantic==2.10.4
tavily-python==0.5.0
duckduckgo-search==8.1.1